+++ src/main/java/chess/domain/board/InitialPieces.java
@@ -1,16 +1,16 @@
-package chess.board;
+package chess.domain.board;
 
-import chess.piece.Bishop;
-import chess.piece.Empty;
-import chess.piece.King;
-import chess.piece.Knight;
-import chess.piece.Pawn;
-import chess.piece.Piece;
-import chess.piece.Queen;
-import chess.piece.Rook;
-import chess.piece.Team;
-import chess.piece.coordinate.Column;
-import chess.piece.coordinate.Coordinate;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.Empty;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
+import chess.domain.piece.Team;
+import chess.domain.piece.coordinate.Column;
+import chess.domain.piece.coordinate.Coordinate;
 import java.util.List;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
