+++ src/main/java/chess/domain/piece/Empty.java
@@ -1,6 +1,6 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.piece.coordinate.Coordinate;
+import chess.domain.piece.coordinate.Coordinate;
 
 public class Empty extends Piece {
     public Empty(Team team, Coordinate coordinate) {
