+++ src/main/java/chess/domain/piece/PieceMatcher.java
@@ -1,6 +1,6 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.piece.coordinate.Coordinate;
+import chess.domain.piece.coordinate.Coordinate;
 
 import java.util.Arrays;
 import java.util.function.BiFunction;
