+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,6 +1,6 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.piece.coordinate.Coordinate;
+import chess.domain.piece.coordinate.Coordinate;
 
 public class Queen extends Piece {
     public Queen(Team team, Coordinate coordinate) {
