+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,26 +1,25 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.piece.coordinate.Coordinate;
+import chess.domain.piece.coordinate.Coordinate;
 
-public class Bishop extends Piece {
-    public Bishop(Team team, Coordinate coordinate) {
+public class Rook extends Piece {
+    public Rook(Team team, Coordinate coordinate) {
         super(team, coordinate);
     }
     
-    
     @Override
     public SymbolMatcher symbol() {
-        return SymbolMatcher.BISHOP;
+        return SymbolMatcher.ROOK;
     }
     
     @Override
     public boolean isMovable(Piece targetPiece) {
         int rowDistance = calculateRowOrColumnDistance(targetPiece, ROW_INDEX);
         int columnDistance = calculateRowOrColumnDistance(targetPiece, COLUMN_INDEX);
-        return isBishopMovable(targetPiece, rowDistance, columnDistance);
+        return isRookMovable(targetPiece, rowDistance, columnDistance);
     }
     
-    private boolean isBishopMovable(Piece targetPiece, int rowDistance, int columnDistance) {
+    private boolean isRookMovable(Piece targetPiece, int rowDistance, int columnDistance) {
         if (isOutOfMovementRadius(rowDistance, columnDistance)) {
             return false;
         }
@@ -29,10 +28,10 @@ public class Bishop extends Piece {
     }
     
     private boolean isOutOfMovementRadius(int rowDistance, int columnDistance) {
-        return isBothZero(rowDistance, columnDistance) || isBothDifferent(rowDistance, columnDistance);
+        return isBothZero(rowDistance, columnDistance) || isBothNotZero(rowDistance, columnDistance);
     }
     
-    private boolean isBothDifferent(int rowDistance, int columnDistance) {
-        return rowDistance != columnDistance;
+    private boolean isBothNotZero(int rowDistance, int columnDistance) {
+        return rowDistance != 0 && columnDistance != 0;
     }
 }
