+++ src/main/java/chess/domain/piece/coordinate/Row.java
@@ -1,4 +1,4 @@
-package chess.piece.coordinate;
+package chess.domain.piece.coordinate;
 
 import java.util.Arrays;
