+++ src/test/java/chess/board/RowPiecesTest.java
@@ -1,13 +1,11 @@
 package chess.board;
 
-import chess.piece.coordinate.Coordinate;
-import org.assertj.core.api.Assertions;
+import chess.domain.board.RowPieces;
+import chess.domain.piece.coordinate.Coordinate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
