+++ src/test/java/chess/piece/TeamTest.java
@@ -1,6 +1,6 @@
 package chess.piece;
 
-import chess.piece.Team;
+import chess.domain.piece.Team;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.params.ParameterizedTest;
