+++ src/test/java/chess/piece/coordinate/ColumnTest.java
@@ -2,6 +2,7 @@ package chess.piece.coordinate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.piece.coordinate.Column;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
