+++ src/test/java/chess/piece/coordinate/CoordinateTest.java
@@ -1,5 +1,6 @@
 package chess.piece.coordinate;
 
+import chess.domain.piece.coordinate.Coordinate;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
