+++ src/test/java/chess/piece/coordinate/RowTest.java
@@ -1,9 +1,8 @@
 package chess.piece.coordinate;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
-import org.assertj.core.api.Assertions;
+import chess.domain.piece.coordinate.Row;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
